% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_hclust.R
\name{fast_hclust_R}
\alias{fast_hclust_R}
\title{Clustering hiérarchique optimisé (appel à Rcpp)}
\usage{
fast_hclust_R(dist_matrix, method = "single", dendrogramme = TRUE)
}
\arguments{
\item{dist_matrix}{Matrice de distance symétrique}

\item{method}{"single", "complete", "average"}

\item{dendrogramme}{TRUE pour afficher}
}
\value{
Un objet hclust
}
\description{
Clustering hiérarchique optimisé (appel à Rcpp)
}
